requirements.txt
# ----------------------
# Flask
# ----------------------
flask
flask-restful
sqlalchemy

# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gevent
gunicorn

# ----------------------
# JWT
# ----------------------
pyjwt>=1.0.1
cryptography>=1.7.1
